Finalized Arduino Software
Fixed Typos in const names.
Added additional Serial.println()'s for readability.
Added a check for wifi connection, and retrying to connect each loop, if failed.
Added Debug if section around pump activation.
Added support for Wifi connected Diode on port 12 - which doesn't appear to be working “¯\_(ツ)_/¯
Created a template Secrets.h, to indicate the existence of a hidden file.